hipotecar y deshipotecar implementados

+++ src/main/java/com/tallerwebi/dominio/RepositorioPartidaUsuarioPropiedad.java
@@ -3,6 +3,6 @@ package com.tallerwebi.dominio;
 public interface RepositorioPartidaUsuarioPropiedad {
     public void crearPartidaUsuarioPropiedad(PartidaUsuarioPropiedad pup);
     public void eliminarPartidaUsuarioPropiedadPorJugadorYPropiedad(PartidaUsuario pu , Propiedad p);
-    public PartidaUsuarioPropiedad obtenerPartidaUsuarioPropiedadPorId(Long id);
+    public PartidaUsuarioPropiedad obtenerPartidaUsuarioPropiedadPorId(Integer id);
     public void actualizarPartidaUsuarioPropiedad(PartidaUsuarioPropiedad pup);
 }
